Test for inversional symmetry

Is the pc-set inversionally symmetrical? That is, does it map onto itself under \(T_n I\) for some appropriate \(n\)? This is evaluated by asking whether, for some appropriate rotations, the step-interval series of the given set is equal to the step-interval series of the set's inversion. This is designed to work for sets in continuous pc-space, not just integers mod k. As usual for working with real values, this depends on your rounding tolerance (rounder).

Usage

isym(set, edo = 12, rounder = 10)

Arguments

set

Numeric vector of pitch-classes in the set

edo

Number of unit steps in an octave. Defaults to 12.

rounder

Numeric (expected integer) representing tolerance: number of decimal places to round to.

Value

TRUE if the set is inversionally symmetrical, FALSE otherwise

Details

Note that this calculates abstract pitch-class symmetry, not potential symmetry in pitch space. (See the second example.)

Examples

#### Mod 12
isym(c(0, 1, 5, 8))
#> [1] TRUE
isym(c(0, 2, 4, 8))
#> [1] TRUE

#### Continuous Values
qcm_fifth <- meantone_fifth()
qcm_dia <- sort(((0:6)*qcm_fifth)%%12)
just_dia <- 12 * log2(c(1, 9/8, 5/4, 4/3, 3/2, 5/3, 15/8))
isym(qcm_dia)
#> [1] TRUE
isym(just_dia)
#> [1] FALSE

#### Rounding matters:
isym(qcm_dia, rounder=15)
#> [1] FALSE