Distances between continuous transpositions of a set

One way to think about the voice-leading potential of a set is to consider all the ways that it can move to transpositions of itself. For instance, the major triad's closest transpositions are \(T_4\) and \(T_8\) while its most distant transposition is \(T_6\), and potentially also \(T_{\pm 2}\) depending on the distance metric you use. For the major triad restricted to 12-tone equal temperament, this set of relationships is well modeled by Richard Cohn's discussion of Douthett & Steinbach's "Cube Dance" in Audacious Euphony (102-106). The behavior of other sets is not always what you might expect based on the model of tertian sonorities. For instance, the trichord (027) has different minimal neighbors depending on the metric chosen: its nearest neighbors are \(T_{\pm 4}\) under the Euclidean metric but \(T_{\pm 5}\) under the taxicab metric.

Usage

tndists(set, method = c("taxicab", "euclidean"), subdivide = 100, edo = 12)