Simplify a scale

simplify_scale description

Usage

simplify_scale(
  set,
  start_zero = TRUE,
  ineqmat = NULL,
  scales = NULL,
  signvector_list = NULL,
  adjlist = NULL,
  method = c("euclidean", "taxicab", "chebyshev", "hamming"),
  display_digits = 2,
  edo = 12,
  rounder = 10
)

best_simplification(set, ...)

Arguments

set

Numeric vector of pitch-classes in the set

start_zero

Boolean: should the result be transposed so that its pitch initial is zero? Defaults to TRUE.

ineqmat

Defaults to NULL, in which case the function assumes you want to use the standard hyperplane arrangement of MCT. But you can enter a different matrix that contains normal vectors for any hyperplane arrangement in the same format as the standard ineqmats.

scales

List of scales representing the faces of your hyperplane arrangement. Defaults to NULL in which case the function looks for representative_scales in the global environment.

signvector_list

A list of signvectors to use as the reference by which colornum assigns a value. Defaults to NULL and will attempt to use representative_signvectors, which needs to be downloaded and assigned separately from the package musicMCT.

adjlist

Adjacency list structed in the same way as color_adjacencies. Defaults to NULL in which case the function looks for color_adjacencies in the global environment.

method

What distance metric should be used? Defaults to "euclidean" (unlike most functions with a method parameter in musicMCT) but can be "taxicab", "chebyshev", or "hamming".

display_digits

Integer: how many digits to display when naming any non-integral interval sizes. Defaults to 2.

edo

Number of unit steps in an octave. Defaults to 12.

rounder

Numeric (expected integer), defaults to 10: number of decimal places to round to when testing for equality.

...

Other arguments to be passed from best_simplification() to simplify_scale().

Value

A complicated matrix

Examples

if (FALSE) { # \dontrun{
  simplify_scale(c(0,5))
} # }