Specify a scale's step pattern with a sign vector

Rather than calculate the full sign vector from the "modal color" hyperplane arrangement, sometimes it's advantageous to use a signvector that reflects only the pairwise comparisons on a scale's steps. This function does that.

Usage

step_signvector(set, ineqmat = NULL, edo = 12, rounder = 10)

Arguments

set

Numeric vector of pitch-classes in the set

ineqmat

Defaults to NULL, in which case the function assumes you want to use the standard hyperplane arrangement of MCT. But you can enter a different matrix that contains normal vectors for any hyperplane arrangement in the same format as the standard ineqmats.

edo

Number of unit steps in an octave. Defaults to 12.

rounder

Numeric (expected integer) representing tolerance: number of decimal places to round to.

Value

A vectors of signs, -1, 0, and 1, corresponding to the step-related hyperplanes in the defined ineqmat.

Examples

step_signvector(sc(7,35)) # Half the length of a full sign vector for heptachords:
#>  [1] -1 -1  0  0  1  1 -1  0  0 -1 -1  0  0 -1  0 -1  0  0 -1  0  0
signvector(sc(7,35))
#>  [1] -1 -1  0  0  1  1 -1  0  0 -1 -1  0  0 -1  0 -1  0  0 -1  0  0  0  0  1 -1
#> [26]  0 -1 -1  0 -1 -1  1  0  0  1  0 -1 -1  0  0  0  0