Create a Scala tuning file from a given scale

You mean you don't want to play around in R forever? This function lets you export any scale you've defined in R as a .scl tuning file for use in Scala or by any synth that can read .scl files. Will write to your working directory.

In addition to saving the necessary tuning data, the function will attempt to add as comments extra information that can be derived from MCT functions, like the color number, degrees of freedom, number of sign-vector zeroes, etc.

Usage

writeSCL(x, filename, period = 2, ineqmat = NULL, edo = 12, rounder = 10)

Arguments

x

Numeric vector: the scale to export

filename

String (in quotation marks): what to name your Scala file. Defaults to using the name of x as the file name if you enter nothing.

period

The frequency ratio at which your scale repeats; defaults to 2 which indicates an octave-repeating scale.

ineqmat

Defaults to NULL, in which case the function assumes you want to use the standard hyperplane arrangement of MCT. But you can enter a different matrix that contains normal vectors for any hyperplane arrangement in the same format as the standard ineqmats.

edo

Number of unit steps in an octave. Defaults to 12.

rounder

Numeric (expected integer), defaults to 10: number of decimal places to round to when testing for equality.

Value

Invisible NULL

Examples

neat_pentachord <- convert(c(0,1,4,9,11),15,12)
if (FALSE) { # \dontrun{
  writeSCL(neat_pentachord, "neat_pentachord.scl")
} # }